refactor: move CarbonAds component to home dir
